Move logic to SwapEditorButton.vue
